refactor(pnm-cli): extract subcommands into dedicated modules
main.rs (2,951 lines) is now a thin dispatch table (439 lines)
delegating to per-subcommand modules under src/commands/{acl,
audit,auth,auth_credential,backup,bootstrap,config,contexts,
did_templates,health,keys,services,setup,vta,webvh}. Clap struct
definitions + parser-adjacent helpers (requires_auth,
is_online_template_cmd, retired_mediator_redirect, print_banner,
install_force_exit_handler) moved to a new src/cli.rs.

Each subcommand handler is a `pub(crate) async fn run(...)` (or
per-action functions for nested subcommands like services and
bootstrap that need to split offline/authed dispatch). The
inline `cmd_health`, `cmd_restart`, `cmd_backup_export`, and
`cmd_backup_import` helpers moved into commands/health.rs,
commands/vta.rs, and commands/backup.rs respectively. No
behavior change.

Refs #7 in the workspace audit.
